Organized backend folder and created apple2_scraper.py
scrapers and csvs are now organized into their own folders in /backend. I created a new scraper for games on the Apple II computer and stored the results in /backend/csvs.

I also greatly improved my code that finds the genre for a game. It proved to add a few minutes to run my scripts, but not a big deal when you only have to scrape once